Middle Save > Change Billboard System > Change Player OBJ 3D to 2D (Billboard) > Add Some GUI Menu > Make Game Screen Manager > Change Shader, Light Manager Function For 2D OBJ In 3D World

+++ Program_Files/Draw_Resource/Effect.h
@@ -11,7 +11,7 @@
 void Effect_Initialize();
 void Effect_Finalize();
 
-void Effect_Update(double elapsed_time);
+void Effect_Update(float elapsed_time);
 void Effect_Draw();
 
 void Effect_Create(const DirectX::XMFLOAT2& position);
